FEniCS backend: Move symbolic expression code
