+++ tlm_adjoint/fenics/functions.py
@@ -2,38 +2,26 @@
 and Dirichlet boundary conditions.
 """
 
-from .backend import (
-    FunctionSpace, TensorFunctionSpace, TestFunction, TrialFunction,
-    VectorFunctionSpace, backend_Constant, backend_Function,
-    backend_ScalarType, cpp_Constant)
+from .backend import backend_Constant, backend_ScalarType
 from ..interface import (
-    SpaceInterface, VariableInterface, VariableStateChangeError,
-    add_replacement_interface, is_var, manager_disabled, space_comm, var_comm,
-    var_dtype, var_is_cached, var_is_replacement, var_is_static,
-    var_linf_norm, var_lock_state, var_replacement, var_scalar_value,
-    var_space, var_space_type)
+    SpaceInterface, VariableInterface, space_comm, var_comm, var_dtype,
+    var_is_cached, var_is_static, var_linf_norm, var_lock_state,
+    var_scalar_value, var_space_type)
 
 from ..caches import Caches
 
-from collections.abc import Sequence
-import functools
+from .expr import Replacement, Zero
+
 import numbers
 import numpy as np
-try:
-    import ufl_legacy as ufl
-except ImportError:
-    import ufl
 import weakref
 
 __all__ = \
     [
         "Constant",
 
-        "Zero",
         "ZeroConstant",
-        "eliminate_zeros",
 
-        "Replacement",
         "ReplacementConstant",
         "ReplacementFunction",
         "ReplacementZeroConstant",
@@ -298,16 +286,6 @@ class Constant(backend_Constant):
         self._tlm_adjoint__var_interface_attrs.d_setitem("cache", cache)
 
 
-class Zero:
-    """Mixin for defining a zero-valued variable. Used for zero-valued
-    variables for which UFL zero elimination should not be applied.
-    """
-
-    def _tlm_adjoint__var_interface_update_state(self):
-        raise VariableStateChangeError("Cannot call _update_state interface "
-                                       "of Zero")
-
-
 class ZeroConstant(Constant, Zero):
     """A :class:`.Constant` which is flagged as having a value of zero.
 
@@ -323,149 +301,6 @@ class ZeroConstant(Constant, Zero):
         var_lock_state(self)
         if var_linf_norm(self) != 0.0:
             raise RuntimeError("ZeroConstant is not zero-valued")
-
-
-def form_cached(key):
-    def wrapper(fn):
-        @functools.wraps(fn)
-        def wrapped(expr, *args, **kwargs):
-            if isinstance(expr, ufl.classes.Form) and key in expr._cache:
-                value = expr._cache[key]
-            else:
-                value = fn(expr, *args, **kwargs)
-                if isinstance(expr, ufl.classes.Form):
-                    assert key not in expr._cache
-                    expr._cache[key] = value
-            return value
-        return wrapped
-    return wrapper
-
-
-@form_cached("_tlm_adjoint__extract_coefficients")
-def extract_coefficients(expr):
-    def as_ufl(expr):
-        if isinstance(expr, (ufl.classes.Expr, numbers.Complex)):
-            return ufl.as_ufl(expr)
-        elif isinstance(expr, ufl.classes.Form):
-            return expr
-        elif isinstance(expr, Sequence):
-            return ufl.as_vector(tuple(map(as_ufl, expr)))
-        else:
-            raise TypeError(f"Unexpected type: {type(expr)}")
-
-    return ufl.algorithms.extract_coefficients(as_ufl(expr))
-
-
-@manager_disabled()
-def is_valid_r0_space(space):
-    e = space.ufl_element()
-    if (e.family(), e.degree()) != ("Real", 0):
-        return False
-    elif len(e.value_shape()) == 0:
-        r = backend_Function(space)
-        r.assign(backend_Constant(-1.0))
-        return (r.vector().max() == -1.0)
-    else:
-        r = backend_Function(space)
-        r_arr = -np.arange(1, np.prod(r.ufl_shape) + 1,
-                           dtype=backend_ScalarType)
-        r_arr.shape = r.ufl_shape
-        r.assign(backend_Constant(r_arr))
-        for i, r_c in enumerate(r.split(deepcopy=True)):
-            if r_c.vector().max() != -(i + 1):
-                return False
-        else:
-            return True
-
-
-def r0_space(x):
-    domain = var_space(x)._tlm_adjoint__space_interface_attrs["domain"]
-    domain = domain.ufl_cargo()
-    if not hasattr(domain, "_tlm_adjoint__r0_space"):
-        if len(x.ufl_shape) == 0:
-            space = FunctionSpace(domain, "R", 0)
-        elif len(x.ufl_shape) == 1:
-            dim, = ufl.shape
-            space = VectorFunctionSpace(domain, "R", 0, dim=dim)
-        else:
-            space = TensorFunctionSpace(domain, "R", degree=0,
-                                        shape=x.ufl_shape)
-        if not is_valid_r0_space(space):
-            raise RuntimeError("Invalid space")
-        domain._tlm_adjoint__r0_space = space
-    return domain._tlm_adjoint__r0_space
-
-
-def derivative_space(x):
-    if isinstance(x, (backend_Constant, ReplacementConstant)):
-        return r0_space(x)
-    else:
-        return var_space(x)
-
-
-def derivative(expr, x, argument=None, *,
-               enable_automatic_argument=True):
-    expr_arguments = ufl.algorithms.extract_arguments(expr)
-    arity = len(expr_arguments)
-
-    if argument is None and enable_automatic_argument:
-        Argument = {0: TestFunction, 1: TrialFunction}[arity]
-        argument = Argument(derivative_space(x))
-
-    for expr_argument in expr_arguments:
-        if expr_argument.number() >= arity:
-            raise ValueError("Unexpected argument")
-    if argument is not None:
-        for expr_argument in ufl.algorithms.extract_arguments(argument):
-            if expr_argument.number() < arity:
-                raise ValueError("Invalid argument")
-
-    return ufl.derivative(expr, x, argument=argument)
-
-
-def expr_zero(expr):
-    if isinstance(expr, ufl.classes.Form):
-        return ufl.classes.Form([])
-    elif isinstance(expr, ufl.classes.Expr):
-        return ufl.classes.Zero(shape=expr.ufl_shape,
-                                free_indices=expr.ufl_free_indices,
-                                index_dimensions=expr.ufl_index_dimensions)
-    else:
-        raise TypeError(f"Unexpected type: {type(expr)}")
-
-
-@form_cached("_tlm_adjoint__eliminate_zeros")
-def eliminate_zeros(expr):
-    """Apply zero elimination for :class:`.Zero` objects in the supplied
-    :class:`ufl.core.expr.Expr` or :class:`ufl.Form`.
-
-    :arg expr: A :class:`ufl.core.expr.Expr` or :class:`ufl.Form`.
-    :returns: A :class:`ufl.core.expr.Expr` or :class:`ufl.Form` with zero
-        elimination applied. May return `expr`.
-    """
-
-    replace_map = {c: expr_zero(c)
-                   for c in extract_coefficients(expr)
-                   if isinstance(c, Zero)}
-    if len(replace_map) == 0:
-        simplified_expr = expr
-    else:
-        simplified_expr = ufl.replace(expr, replace_map)
-
-    return simplified_expr
-
-
-class Replacement(ufl.classes.Coefficient):
-    """Represents a symbolic variable but with no value.
-    """
-
-    def __init__(self, x, count):
-        super().__init__(var_space(x), count=count)
-        add_replacement_interface(self, x)
-
-
-def new_count():
-    return cpp_Constant(0.0).id()
 
 
 class ReplacementConstant(Replacement):
@@ -494,16 +329,6 @@ class ReplacementZeroFunction(ReplacementFunction, Zero):
     def __init__(self, *args, **kwargs):
         ReplacementFunction.__init__(self, *args, **kwargs)
         Zero.__init__(self)
-
-
-def replaced_form(form):
-    replace_map = {}
-    for c in extract_coefficients(form):
-        if is_var(c) and not var_is_replacement(c):
-            c_rep = var_replacement(c)
-            if c_rep is not c:
-                replace_map[c] = c_rep
-    return ufl.replace(form, replace_map)
 
 
 def define_var_alias(x, parent, *, key):
